addresses

+++ coap-sensors/resources/res-scale.c
@@ -6,7 +6,7 @@
 #include <string.h>
 
 // Scale sensor state
-static float scale_value = 32.0; // Initial weight value
+static int scale_value = 32; // Initial weight value
 
 // Conversion functions for the scale sensor
 static void scale_value_to_string(char *buffer, size_t size, void *state) {
@@ -14,12 +14,12 @@ static void scale_value_to_string(char *buffer, size_t size, void *state) {
 }
 
 static void scale_value_update_state(const char *payload, void *state) {
-  *(float *)state += atof(payload); // Convert payload string to float and add to the current value
+  *(int *)state += atoi(payload); // Convert payload string to float and add to the current value
   // Clamp to minimum 0
-    if (*(float *)state < 0.0) {
-        *(float *)state = 0.0;
+    if (*(int *)state < 0) {
+        *(int *)state = 0;
     }
-  printf("Scale sensor value updated to: %.2f\n", *(float *)state);
+  printf("Scale sensor value updated to: %.2d\n", *(int *)state);
 }
 
 // Define the generic sensor structure
